requirements.txt
Package                  Version
------------------------ ------------
Flask                    3.0.3
Werkzeug                 3.0.4
python-dotenv            1.0.1
PyMuPDF                  1.24.10
openai                   1.44.0
langchain                0.2.16
langchain-community      0.2.16
langchain-core           0.2.38
langchain-text-splitters 0.2.4
